+++ app/models/rails_pulse/exception_occurrence.rb
@@ -19,5 +19,9 @@ module RailsPulse
     def self.ransackable_associations(auth_object = nil)
       %w[exception_group]
     end
+
+    def to_breadcrumb
+      occurred_at.getlocal.strftime("%b %d, %Y %l:%M %p")
+    end
   end
 end