+++ lib/rails_pulse/job_run_collector.rb
@@ -43,6 +43,7 @@ module RailsPulse
           ) if job_run
         end
         RailsPulse::ExceptionCaptureService.capture(error, environment: Rails.env.to_s)
+        RequestStore.store[:rails_pulse_captured_exception] = error
         raise
       ensure
         begin